+++ src/examples/mod.rs
@@ -85,9 +85,12 @@ pub mod basic {
 
     #[test]
     fn run() {
-        use std::mem::size_of;
+        use core::mem::size_of;
         assert_eq!(size_of::<VarIndex>(), size_of::<usize>());
-        assert_eq!(size_of::<VarMap<String>>(), size_of::<Vec<String>>());
+        assert_eq!(
+            size_of::<VarMap<alloc::string::String>>(),
+            size_of::<alloc::vec::Vec<alloc::string::String>>()
+        );
 
         let mut var_values = VarMap::with_capacity(3);
         let v_0 = var_values.push(7);
@@ -117,19 +120,19 @@ pub mod basic {
         assert_eq! { iter.next(), Some(&11) }
         assert_eq! { iter.next(), None      }
 
-        let mut check = vec![11, 3, 7];
+        let mut check = alloc::vec![11, 3, 7];
         for val in &var_values {
             assert_eq! { *val, check.pop().unwrap() }
         }
 
-        let mut check = vec![(v_2, 11), (v_1, 3), (v_0, 7)];
+        let mut check = alloc::vec![(v_2, 11), (v_1, 3), (v_0, 7)];
         for (idx, val) in var_values.index_iter() {
             let (i, v) = check.pop().unwrap();
             assert_eq! {  idx, i }
             assert_eq! { *val, v }
         }
 
-        let mut check = vec![11, 3, 7];
+        let mut check = alloc::vec![11, 3, 7];
         for idx in var_values.indices() {
             assert_eq! { var_values[idx], check.pop().unwrap() }
         }